requirements.txt
httpx==0.23.1
python-telegram-bot==13.15
telegram-send==0.34
python-decouple==3.7
h2==4.1.0

